leading_lights/requirements.txt
# Leading Lights Package Dependencies

# Core
requests>=2.28.0
pyyaml>=6.0
pandas>=1.5.0

# Optional: LinkedIn scraping (paid)
# proxycurl-py  # Uncomment if using Proxycurl API
